% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumHEM.R
\name{SumHEM}
\alias{SumHEM}
\title{Summary-level heteroscedastic effects model (SumHEM)}
\usage{
SumHEM(df_gwas, df_map, LD_path, h2input = "", ws = 500, NCORES = 1)
}
\arguments{
\item{df_gwas}{A data frame including GWAS summary statistics of genetic variants for a trait. 
The input data frame should include following columns: 
rsid, SNP ID; chr, chromosome where the SNP is; pos, position of the SNP; 
beta, estimate of marginal effect in GWAS; se, standard error of the estimates of marginal effects in GWAS; N, sample size;}

\item{df_map}{A data frame including SNP information of whole-genome linkage disequilibrium (LD) reference panel. 
The information should be totally matched with the SNP including in LD files and the input data frame should include following columns: 
rsid, SNP ID; chr, chromosome where the SNP is; pos, position of the SNP;}

\item{LD_path}{Path to the directory where LD information is stored. 
The LD matrix should be separated by chromosomes and stored in 'dsCMatrix' class as a 'rds' object in each LD file. 
The filenames should be 'LD_chr'+chromosome+'.rds'}

\item{h2input}{Heritability of the trait estimated by external software. 
The input numeric value should be between 0 and 1. 
If missing, the value will be computed in progress.}

\item{ws}{The number of SNP in a window for estimation.}

\item{NCORES}{The number of cores using for parallel computation.}
}
\value{

}
\description{
The function returns the estimates of SNP effects of one trait based on GWAS summary statistics.
}
\note{
Users can download the LD correlation matrices for European ancestry population, authored by Florian Privé. 
The download link can be found at https://figshare.com/articles/dataset/European_LD_reference/13034123
This is the LD reference (correlations between pairs of genetic variants) for 1,054,330 HapMap3 variants based on 362,320 European individuals of the UK biobank.
}
\examples{
\dontrun{
## Load example GWAS summary statistics for standing height
df_gwas <- readRDS("SumHEM/data/gwas.height.rds")
## Load SNP information of LD reference panel
df_map <- readRDS("SumHEM/data/map.rds")

## Setting for parameters
ws <- 500 # the number of SNPs in a window for estimation
NCORES <- 10 # the number of cores using in parallel computation
LD_path <- "SumHEM/LD_EUR" # path of folder with LD files
h2input <- 0.4780 # heritability estimated by an external software (optional)

## Estimate SNP effects via SumHEM
df_est <- SumHEM(
  df_gwas = df_gwas, df_map = df_map,
  h2input = h2input, # if missing, it will be computed during the process
  ws = ws, NCORES = NCORES, LD_path = LD_path
)

## View the results
str(df_est)

## Summaries the inferred heritability
df <- df_est \%>\% distinct(chr,winID, .keep_all = TRUE)
sum(df$h2_SumRR) # inferred heritability of SumRR
sum(df$h2_SumHEM) # inferred heritability of SumHEM
sqrt(sum(df$h2var_SumHEM)) # se of SumHEM-inferred heritability

}
}
\references{
...
}
\seealso{
SumHEM tutorial on GitHub page: https://github.com/yueyao0/SumHEM
}
\author{
Yue Yao, Wenzhuo Lin, Xia Shen
}
